Rename module
Futurely, other kind of transaction may exist, so this transaction module was specified for bulking.
