+++ vitae/infra/database/bulk_transactions.py
